requirements.txt
asgiref==3.2.7
cycler==0.10.0
Django==3.0.5
django-crispy-forms==1.9.0
joblib==0.14.1
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.9.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
xgboost==1.0.2
